[feat][12] 주문-결제 로직 구현. 예치금 관련 로직은 주석처리

+++ payment/src/main/java/com/ll/payment/model/vo/TossPaymentRequest.java
@@ -3,6 +3,6 @@ package com.ll.payment.model.vo;
 public record TossPaymentRequest(
         String paymentKey,
         String orderId,
-        Long amount
+        int amount
 ) {
 }
